backend/requirements.txt
# Core Dependencies
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Database & Supabase
supabase==2.3.4
sqlalchemy==2.0.25
asyncpg==0.29.0

# OpenRouter & AI
httpx==0.26.0
pydantic==2.5.3
python-dotenv==1.0.0

# Cache & Performance
redis==5.0.1
aiocache==0.12.3

# Security & Validation
email-validator==2.1.0
python-decouple==3.8

# Monitoring & Logging
loguru==0.7.2
prometheus-client==0.19.0

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
httpx==0.26.0